# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs DTU_DZ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/chuong/data": {
            "get": {
                "tags": [
                    "Chương"
                ],
                "summary": "Chuong Data",
                "description": "Get Data Chuong",
                "operationId": "2a58104f72ea1b54360c1ef04510e721",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            }
        },
        "/api/admin/chuong/doi-trang-thai": {
            "post": {
                "tags": [
                    "Chương"
                ],
                "summary": "Doi Trang Thai Chuong",
                "description": "",
                "operationId": "c48a327249f98d5188bf1d396952d67a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/admin/chuong/update": {
            "post": {
                "tags": [
                    "Chương"
                ],
                "summary": "Update Chuong",
                "description": "",
                "operationId": "b024356de5969cdcce79dc714cf08144",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "ten_chuong",
                                    "so_thu_tu",
                                    "id_de_muc",
                                    "is_open"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "ten_chuong": {
                                        "type": "text"
                                    },
                                    "so_thu_tu": {
                                        "type": "number"
                                    },
                                    "id_de_muc": {
                                        "type": "number"
                                    },
                                    "is_open": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}